Moving the new fixture to pfcwd_helper.py
