+++ tests/pfcwd/test_pfcwd_timer_accuracy.py
@@ -1,12 +1,12 @@
 import logging
 import pytest
 import time
-import random
 
 from tests.common.fixtures.conn_graph_facts import enum_fanout_graph_facts      # noqa F401
 from tests.common.helpers.assertions import pytest_assert
 from tests.common.helpers.pfc_storm import PFCStorm
-from .files.pfcwd_helper import start_wd_on_ports
+from .files.pfcwd_helper import start_wd_on_ports, start_background_traffic     # noqa F401
+
 from tests.common.plugins.loganalyzer import DisableLogrotateCronContext
 
 
@@ -15,6 +15,13 @@ pytestmark = [
 ]
 
 logger = logging.getLogger(__name__)
+
+
+@pytest.fixture(scope="class")
+def pfc_queue_idx(pfcwd_timer_setup_restore):
+    # This is used by the common code, this needs to be defined
+    # before using start_background_traffic() fixture.
+    yield pfcwd_timer_setup_restore['storm_handle'].pfc_queue_idx
 
 
 @pytest.fixture(scope='module', autouse=True)
@@ -155,110 +162,7 @@ def set_storm_params(dut, fanout_info, fanout, peer_params):
     return storm_handle
 
 
-@pytest.fixture(scope='class', autouse=True)
-def start_background_traffic(
-        duthosts,
-        enum_rand_one_per_hwsku_frontend_hostname,
-        pfcwd_timer_setup_restore,
-        setup_pfc_test,
-        copy_ptftests_directory,
-        ptfhost,
-        tbinfo
-        ):
-    """
-       This fixutre is to start a background traffic during
-       the test. This will start a continuous traffic flow from PTF
-       exiting the test port.
-    """
-    if duthosts[enum_rand_one_per_hwsku_frontend_hostname].facts['asic_type'] != "cisco-8000":
-        yield
-        return
-
-    # This is needed only for cisco-8000
-    program_name = "pfcwd_background_traffic"
-    dut = duthosts[enum_rand_one_per_hwsku_frontend_hostname]
-    dst_dut_intf = list(setup_pfc_test['test_ports'].keys())[0]
-    mg_facts = dut.get_extended_minigraph_facts(tbinfo)
-    vlan_ports = []
-    for vlan in mg_facts['minigraph_vlans'].keys():
-        vlan_ports.extend(mg_facts['minigraph_vlans'][vlan]['members'])
-    all_ip_intfs = mg_facts['minigraph_interfaces'] + mg_facts['minigraph_portchannel_interfaces']
-    non_vlan_ports = set(list(setup_pfc_test['test_ports'])) - set(vlan_ports) - set([dst_dut_intf])
-    src_dut_intf = random.choice(list(non_vlan_ports))
-    dest_mac = dut.get_dut_iface_mac(src_dut_intf)
-    # Find out if the selected port is a lag member
-    # If so, we need to use the neighbor address of the portchannel.
-    # else, we need the neighbor address of the interface itself.
-    required_intf = dst_dut_intf
-    for intf in mg_facts['minigraph_portchannels']:
-        if dst_dut_intf in mg_facts['minigraph_portchannels'][intf]['members']:
-            required_intf = intf
-            break
-    # At this point, required_intf is either a portchannel or Ethernet port.
-    # It should have a neibhor address or it is an error.
-    dst_ip_addr = None
-    for intf_obj in all_ip_intfs:
-        if intf_obj['attachto'] == required_intf:
-            dst_ip_addr = intf_obj['peer_addr']
-            break
-    if dst_ip_addr is None:
-        raise RuntimeError("Couldnot find the neighbor address for intf:{}".format(required_intf))
-    ptf_src_port = mg_facts['minigraph_ptf_indices'][src_dut_intf]
-    ptf_dst_port = mg_facts['minigraph_ptf_indices'][dst_dut_intf]
-    extra_vars = {
-        f'{program_name}_args':
-            'dest_mac=u"{}";dst_ip_addr={};ptf_src_port={};ptf_dst_port={};pfc_queue_idx={}'.format(
-                dest_mac,
-                dst_ip_addr,
-                ptf_src_port,
-                ptf_dst_port,
-                pfcwd_timer_setup_restore['storm_handle'].pfc_queue_idx
-                )}
-    try:
-        ptfhost.command('supervisorctl stop {}'.format(program_name))
-    except BaseException:
-        pass
-
-    ptfhost.host.options["variable_manager"].extra_vars.update(extra_vars)
-    script_args = \
-        '''dest_mac=u"{}";dst_ip_addr="{}";ptf_src_port={};ptf_dst_port={};pfc_queue_idx={}'''.format(
-                dest_mac,
-                dst_ip_addr,
-                ptf_src_port,
-                ptf_dst_port,
-                pfcwd_timer_setup_restore['storm_handle'].pfc_queue_idx)
-    supervisor_conf_content = ('''
-[program:{program_name}]
-command=/root/env-python3/bin/ptf --test-dir /root/ptftests/py3 {program_name}.BG_pkt_sender'''
-                               ''' --platform-dir /root/ptftests/ -t'''
-                               ''' '{script_args}' --relax  --platform remote
-process_name={program_name}
-stdout_logfile=/tmp/{program_name}.out.log
-stderr_logfile=/tmp/{program_name}.err.log
-redirect_stderr=false
-autostart=false
-autorestart=true
-startsecs=1
-numprocs=1
-'''.format(script_args=script_args, program_name=program_name))
-    ptfhost.copy(
-        content=supervisor_conf_content,
-        dest=f'/etc/supervisor/conf.d/{program_name}.conf')
-
-    ptfhost.command('supervisorctl reread')
-    ptfhost.command('supervisorctl update')
-    ptfhost.command(f'supervisorctl start {program_name}')
-
-    yield
-
-    try:
-        ptfhost.command(f'supervisorctl stop {program_name}')
-    except BaseException:
-        pass
-    ptfhost.command(f'supervisorctl remove {program_name}')
-
-
-@pytest.mark.usefixtures('pfcwd_timer_setup_restore')
+@pytest.mark.usefixtures('pfcwd_timer_setup_restore', 'start_background_traffic')
 class TestPfcwdAllTimer(object):
     """ PFCwd timer test class """
     def run_test(self):
